Impl cpu instructions

+++ src/cpu/cpu_bus.cpp
@@ -1,5 +1,5 @@
 #include "../common.hpp"
 
-uint8_t CPU6502Bus::read(uint8_t address) {
+uint8_t CPU6502Bus::read(uint16_t address) {
    return this->ram->read(address);
 }
